btc_signal_dashboard.py

+++ btc_signal_dashboard.py
@@ -27,6 +27,8 @@ def get_ohlcv_from_yfinance(symbol, period='90d', interval='1d'):
         df = yf.download(symbol, period=period, interval=interval, progress=False)
         df = df[['Close', 'Volume']].rename(columns={'Close': 'close', 'Volume': 'volume'})
         df = df.reset_index(drop=True)
+        df['close'] = df['close'].astype(float)
+        df['volume'] = df['volume'].astype(float)
         return df
     except Exception as e:
         st.error(f"{symbol} veri çekme hatası: {str(e)}")
@@ -60,14 +62,14 @@ def calculate_signal(df):
     df['VolMA20'] = df['volume'].rolling(window=20).mean()
 
     try:
-        close = df['close'].iloc[-1]
-        ema20 = df['EMA20'].iloc[-1]
-        ema50 = df['EMA50'].iloc[-1]
-        macd = df['MACD'].iloc[-1]
-        macd_signal = df['MACDSignal'].iloc[-1]
-        rsi = df['RSI'].iloc[-1]
-        volume = df['volume'].iloc[-1]
-        volma = df['VolMA20'].iloc[-1]
+        close = float(df['close'].iloc[-1])
+        ema20 = float(df['EMA20'].iloc[-1])
+        ema50 = float(df['EMA50'].iloc[-1])
+        macd = float(df['MACD'].iloc[-1])
+        macd_signal = float(df['MACDSignal'].iloc[-1])
+        rsi = float(df['RSI'].iloc[-1])
+        volume = float(df['volume'].iloc[-1])
+        volma = float(df['VolMA20'].iloc[-1])
 
         if close > ema20 > ema50 and macd > macd_signal and volume > volma and 40 < rsi < 70:
             return 'LONG'
